count: add conservative updates
A 'conservative update' increments the counters to at most the minimum
counter plus the increment value (in this case '1').  This reduces the
amount of over-estimation in the case of hash collisions.  The
downside is that items can no longer be deleted from the sketch as the
update is now lossy -- we can't be sure how much to decrement the
counter by on element removal.