count: inline hash values calculations instead of returning slice
The previous implementation of hashn() return a slice of ints, requiring
an allocation.  This made sense when the hashing algorithms were more
complex, but since they're just linear combinations, we can do that in
the caller instead.  This removes an allocation which ends up doubling
the speed of the hash function: from ~1500ns/op to ~750ns/op.

Fixes issue #6.