Fix docs for the hashing stuff.
The full tech report (as opposed to the journal article, which I cited
originally) has more details on their Count-Min construction.  Using the
linear combinations of hash functions is not a drop-in replacement --
there are other requirements for their proof to actually work.  However,
in my testing with

    h := h1 + i*h2

I've never had any problems, and in fact this fixed a bug with the
original hash stuff when I introduced it.

However, since the proof doesn't apply, I'm a bit concerned about this.
Perhaps it's going to fall over in some wacky edge case.  You never can
tell once the proof no longer holds.

I added a TODO to find a better hash construction for the k-pairwise
independent hash function requirement, *without* resorting to have to
generate and expose seeds for the hash functions.  My initial thoughts
are to keep the two hash functions, but instead change the generation
procedure to

   h := h1 ^ h2
   h2 = intmix(h2);

where intmix is a 32-bit integer hash function.